chore(license): update VDP license to ELv2 and MIT (#418)
Because

- our mission is to make AI accessible to everyone. The best way to
achieve this is to make VDP free to use and source available to
everyone, while ensuring we safely create a sustainable business.
Therefore, we chose ELv2 since it is very permissive. On the one hand,
this allows us to provide users with free access to our source code, the
permission to modify it. On the other hand, we've released Instill
Cloud, a fully-managed public cloud service built upon VDP that offers
additional features. With the license, we don't have to worry about our
project being taken by some other companies for monetization.

This commit

- add LICENSE
- update integration test GitHub repository