fix: replace logout link with button (#1762)
Previously we use <Link /> component in the `CETopbarDropdown`, when
user open this dropdown, Nextjs will prefetch the link to pre-render the
target page. But we use the api route path in the Link's href props.
This means when the Nextjs prefetch the page, it's actually calling the
api route, this operates the logout procedure. Hence user need to
re-login, this PR replace the Nextjs Link component with the Button, so
Nextjs won't pre-fetch it anymore