chore(github,toolkit,radix,jsx): houseclean  codebase (#1814)
Because

- the `.github` folder is pretty outdated.
- dashboard complains about `ECharts` import
- Radix UI complains about `ConsoleError: Accessing element.ref was
removed in React 19. ref is now a regular prop. It will be removed from
the JSX Element type in a future release.`
- Fix JSX errors:
```
Error: Received `true` for a non-boolean attribute `jsx`.

If you want to write it to the DOM, pass a string instead: jsx="true" or jsx={value.toString()}.
```

This commit

- addresses the above issues.