fix(console,tookit): fix misleading errors (#1822)
Because

- the previous logic assumed that data will always be a valid JSON
string that can be parsed, but if the API response is null or doesn't
contain the expected structure, JSON.parse(data) will fail or return
null, causing the error when trying to access `.cookieToken`.
- in the profile setting page, the file input’s id was generated via
`uuidv4()` inside a client component. That produces a different value on
the server vs client during SSR, which triggers the hydration mismatch
you saw in the error diff.
- favicon was outdated.

This commit

- addresses the above issues.
- uses `React.useId()` instead, which is SSR-safe and stable across
server and client.
- updates favicon files.